move all activities to a single activity package
