+++ app/src/main/java/com/arctouch/codechallenge/activity/base/BaseActivity.kt
@@ -1,4 +1,4 @@
-package com.arctouch.codechallenge.base
+package com.arctouch.codechallenge.activity.base
 
 import android.support.v7.app.AppCompatActivity
 import com.arctouch.codechallenge.api.TmdbApi
